move some events to adapters
