+++ core/dbt/adapters/sql/impl.py
@@ -2,12 +2,12 @@ import agate
 from typing import Any, Optional, Tuple, Type, List
 
 from dbt.adapters.contracts.connection import Connection, AdapterResponse
+from dbt.adapters.events.types import ColTypeChange, SchemaCreation, SchemaDrop
 from dbt.adapters.exceptions import RelationTypeNullError
 from dbt.adapters.base import BaseAdapter, available
 from dbt.adapters.cache import _make_ref_key_dict
 from dbt.adapters.sql import SQLConnectionManager
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import ColTypeChange, SchemaCreation, SchemaDrop
 
 
 from dbt.adapters.base.relation import BaseRelation
