+++ core/dbt/task/generate.py
@@ -30,7 +30,7 @@ from dbt.graph import ResourceTypeSelector
 from dbt.node_types import NodeType
 from dbt.include.global_project import DOCS_INDEX_FILE_PATH
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.adapters.events.types import (
     WriteCatalogFailure,
     CatalogWritten,
     CannotGenerateDocs,
